+++ app/models/kms/master_variant.rb
@@ -1,0 +1,4 @@
+module Kms
+  class MasterVariant < Variant
+  end
+end