+++ db/migrate/20170112133235_add_type_to_variants.rb
@@ -1,0 +1,5 @@
+class AddTypeToVariants < ActiveRecord::Migration[5.0]
+  def change
+    add_column :kms_variants, :type, :string, default: Kms::Variant.name
+  end
+end