+++ lib/kms/catalog/engine.rb
@@ -10,6 +10,10 @@ module Kms
           require_dependency(c)
         end
       end
+      config.generators do |g|
+        g.test_framework :rspec, fixture: false
+        g.fixture_replacement :factory_girl, :dir => 'spec/factories'
+      end
     end
   end
 end