+++ test/catalog_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class CatalogTest < ActiveSupport::TestCase
-  test "truth" do
-    assert_kind_of Module, Catalog
-  end
-end