+++ test/models/option_value_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class OptionValueTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end