+++ test/models/product_option_type_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class ProductOptionTypeTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end