+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.5.0 unreleased
+
+* [added] add MasterVariant class and :master association to Product
+* [changed] better UI - Variants in separate tab, inline variant adding, etc.
+
 ## Version 0.4.0 2016-12-22
 
 * [added] added english translations